src/test/java/com/example/petstore/services/AnimalServiceTests.java
package com.example.petstore.services;

import com.example.petstore.TestUtils.Mocks;
import com.example.petstore.model.AnimalEntity;
import com.example.petstore.repository.AnimalRepository;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.web.servlet.MockMvc;

import java.time.LocalDate;
import java.util.List;

import static org.hamcrest.Matchers.hasSize;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(MockitoExtension.class)
public class AnimalServiceTests {

    @Mock
    private AnimalRepository animalRepository;

    @InjectMocks
    private AnimalService animalService;

    @Test
    public void shouldReturnAnimals_WhenThereAreAnimalsInThePetStore() {
        when(animalRepository.findAll()).thenReturn(Mocks.getAnimals());

        List<AnimalEntity> animals = animalService.getAnimals();

        assertEquals(Mocks.getAnimals(), animals);
    }
}
